# tests/golden/openapi_schema.json
{
  "components": {
    "schemas": {
...
      "DriftJobCreateRequest": {
        "properties": {
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "job_name": {
            "title": "Job Name",
            "type": "string"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "reference_config": {
            "additionalProperties": true,
            "title": "Reference Config",
            "type": "object"
          },
          "schedule": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule"
          },
          "source_config": {
            "additionalProperties": true,
            "title": "Source Config",
            "type": "object"
          },
          "source_type": {
            "default": "local_file",
            "title": "Source Type",
            "type": "string"
          },
          "status": {
            "default": "active",
            "title": "Status",
            "type": "string"
          },
          "thresholds": {
            "additionalProperties": true,
            "title": "Thresholds",
            "type": "object"
          }
        },
        "required": [
          "model_name",
          "job_name"
        ],
        "title": "DriftJobCreateRequest",
        "type": "object"
      },
      "DriftJobResponse": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "job_name": {
            "title": "Job Name",
            "type": "string"
          },
          "last_run_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run At"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "next_run_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Run At"
          },
          "reference_config": {
            "additionalProperties": true,
            "title": "Reference Config",
            "type": "object"
          },
          "schedule": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule"
          },
          "source_config": {
            "additionalProperties": true,
            "title": "Source Config",
            "type": "object"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "thresholds": {
            "additionalProperties": true,
            "title": "Thresholds",
            "type": "object"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "id",
          "model_name",
          "job_name",
          "source_type",
          "status",
          "created_at"
        ],
        "title": "DriftJobResponse",
        "type": "object"
      },
      "DriftJobRunResponse": {
        "properties": {
          "alerts_triggered": {
            "title": "Alerts Triggered",
            "type": "integer"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "features_evaluated": {
            "title": "Features Evaluated",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "max_severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Severity"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "run_metadata": {
            "additionalProperties": true,
            "title": "Run Metadata",
            "type": "object"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "model_name",
          "status",
          "features_evaluated",
          "alerts_triggered",
          "created_at"
        ],
        "title": "DriftJobRunResponse",
        "type": "object"
      },
      "DriftPredictionIngestRequest": {
        "properties": {
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "records": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "minItems": 1,
            "title": "Records",
            "type": "array"
          },
          "source": {
            "default": "api_push",
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "model_name",
          "records"
        ],
        "title": "DriftPredictionIngestRequest",
        "type": "object"
      },
      "DriftPredictionIngestResponse": {
        "properties": {
          "columns": {
            "items": {
              "type": "string"
            },
            "title": "Columns",
            "type": "array"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "rows": {
            "title": "Rows",
            "type": "integer"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "model_name",
          "rows",
          "columns",
          "source"
        ],
        "title": "DriftPredictionIngestResponse",
        "type": "object"
      },
      "DriftReportIn": {
        "properties": {
          "feature": {
            "title": "Feature",
            "type": "string"
          },
          "is_alerting": {
            "title": "Is Alerting",
            "type": "boolean"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "kl_divergence": {
            "title": "Kl Divergence",
            "type": "number"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "psi": {
            "title": "Psi",
            "type": "number"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "model_name",
          "feature",
          "kl_divergence",
          "psi",
          "is_alerting",
          "severity"
        ],
        "title": "DriftReportIn",
        "type": "object"
      },
      "DriftReportResponse": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "feature": {
            "title": "Feature",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_alerting": {
            "title": "Is Alerting",
            "type": "boolean"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "kl_divergence": {
            "title": "Kl Divergence",
            "type": "number"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "psi": {
            "title": "Psi",
            "type": "number"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "id",
          "model_name",
          "feature",
          "kl_divergence",
          "psi",
          "is_alerting",
          "severity",
          "created_at"
        ],
        "title": "DriftReportResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Ares",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/drift/jobs": {
      "get": {
        "operationId": "list_jobs_api_v1_drift_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "model_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Name"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DriftJobResponse"
                  },
                  "title": "Response List Jobs Api V1 Drift Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Jobs",
        "tags": [
          "drift"
        ]
      },
      "post": {
        "operationId": "create_job_api_v1_drift_jobs_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriftJobCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriftJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Job",
        "tags": [
          "drift"
        ]
      }
    },
    "/api/v1/drift/jobs/{job_id}/run": {
      "post": {
        "operationId": "run_job_now_api_v1_drift_jobs__job_id__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriftJobRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Job Now",
        "tags": [
          "drift"
        ]
      }
    },
    "/api/v1/drift/jobs/{job_id}/runs": {
      "get": {
        "operationId": "list_job_runs_api_v1_drift_jobs__job_id__runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DriftJobRunResponse"
                  },
                  "title": "Response List Job Runs Api V1 Drift Jobs  Job Id  Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Job Runs",
        "tags": [
          "drift"
        ]
      }
    },
    "/api/v1/drift/predictions": {
      "post": {
        "operationId": "ingest_predictions_api_v1_drift_predictions_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriftPredictionIngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriftPredictionIngestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Predictions",
        "tags": [
          "drift"
        ]
      }
    },
    "/api/v1/drift/predictions/push": {
      "post": {
        "operationId": "ingest_predictions_push_api_v1_drift_predictions_push_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriftPredictionIngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriftPredictionIngestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Predictions Push",
        "tags": [
          "drift"
        ]
      }
    },
    "/api/v1/drift/reports": {
      "get": {
        "operationId": "list_reports_api_v1_drift_reports_get",
        "parameters": [
          {
            "in": "query",
            "name": "model_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Name"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DriftReportResponse"
                  },
                  "title": "Response List Reports Api V1 Drift Reports Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Reports",
        "tags": [
          "drift"
        ]
      },
      "post": {
        "operationId": "create_report_api_v1_drift_reports_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriftReportIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriftReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Report",
        "tags": [
          "drift"
        ]
      }
    },
...
  }
}